changed form structure

+++ src/Controller/Form/DeleteUser.php
@@ -7,7 +7,6 @@ use Modules\Kernel\File;
 use Modules\Kernel\FileManager;
 use Modules\Kernel\Form;
 use Modules\Kernel\Message;
-use Modules\Kernel\Storage;
 use Modules\Kernel\View;
 use Modules\Router\Router;
 
@@ -17,10 +16,13 @@ class DeleteUser extends Form
 
     function __construct()
     {
+        parent::__construct('POST', [
+            'id' => [
+                'from' => &$_POST,
+                'type' => 'integer',
+            ],
+        ]);
         $this->access('admin');
-
-        $this->user=User::load($_GET['id']);
-        $this->file = File::load($this->user->avatar);
     }
 
     function title(): string
@@ -33,18 +35,20 @@ class DeleteUser extends Form
         return new View('page/delete_user.phtml');
     }
 
-    public function verify(): bool
+    public function verify(&$data)
     {
+        $this->user = User::load($data['id']);
         if (empty($this->user))
-            return $this->error('El usuario no existe');
+            return Message::add('El usuario no existe');
+        if (!empty($this->user->avatar))
+            $this->file = File::load($this->user->avatar);
         return true;
     }
 
-    function submit()
+    function submit(&$data)
     {
-        if(!empty($this->file)){
+        if (isset($this->file) && !empty($this->file))
             FileManager::delete($this->file);
-        }
         User::remove($this->user->id);
         Message::add('Se ha eliminado el usuario');
         return Router::get('/admin/usuarios');
